transfert kernel directory to Xoops/Core/Kernel
