+++ htdocs/xoops_lib/Xoops/Core/Kernel/Handlers/XoopsBlock.php
@@ -1,0 +1,412 @@
+<?php
+/*
+ You may not change or alter any portion of this comment or credits
+ of supporting developers from this source code or any supporting source code
+ which is considered copyrighted (c) material of the original comment or credit authors.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+*/
+
+namespace Xoops\Core\Kernel\Handlers;
+
+use Xoops\Core\Kernel\XoopsObject;
+
+/**
+ * XoopsBlock
+ *
+ * @package   kernel
+ * @author    Kazumi Ono (AKA onokazu) http://www.myweb.ne.jp/, http://jp.xoops.org/
+ * @author    Gregory Mage (AKA Mage)
+ * @author    trabis <lusopoemas@gmail.com>
+ * @copyright 2003-2014 XOOPS Project (http://xoops.org)
+ * @license   GNU GPL 2 or later (http://www.gnu.org/licenses/gpl-2.0.html)
+ * @since     2.0.0
+ */
+class XoopsBlock extends XoopsObject
+{
+    /**
+     * Constructor
+     *
+     * @param int|array $id
+     */
+    public function __construct($id = null)
+    {
+        $this->initVar('bid', XOBJ_DTYPE_INT, null, false);
+        $this->initVar('mid', XOBJ_DTYPE_INT, 0, false);
+        $this->initVar('func_num', XOBJ_DTYPE_INT, 0, false);
+        $this->initVar('options', XOBJ_DTYPE_TXTBOX, null, false, 255);
+        $this->initVar('name', XOBJ_DTYPE_TXTBOX, null, true, 150);
+        //$this->initVar('position', XOBJ_DTYPE_INT, 0, false);
+        $this->initVar('title', XOBJ_DTYPE_TXTBOX, null, false, 150);
+        $this->initVar('content', XOBJ_DTYPE_TXTAREA, null, false);
+        $this->initVar('side', XOBJ_DTYPE_INT, 0, false);
+        $this->initVar('weight', XOBJ_DTYPE_INT, 0, false);
+        $this->initVar('visible', XOBJ_DTYPE_INT, 0, false);
+        // The block_type is in a mess, let's say:
+        // S - generated by system module
+        // M - generated by a non-system module
+        // C - Custom block
+        // D - cloned system/module block
+        // E - cloned custom block, DON'T use it
+        $this->initVar('block_type', XOBJ_DTYPE_OTHER, null, false);
+        $this->initVar('c_type', XOBJ_DTYPE_OTHER, null, false);
+        $this->initVar('isactive', XOBJ_DTYPE_INT, null, false);
+        $this->initVar('dirname', XOBJ_DTYPE_TXTBOX, null, false, 50);
+        $this->initVar('func_file', XOBJ_DTYPE_TXTBOX, null, false, 50);
+        $this->initVar('show_func', XOBJ_DTYPE_TXTBOX, null, false, 50);
+        $this->initVar('edit_func', XOBJ_DTYPE_TXTBOX, null, false, 50);
+        $this->initVar('template', XOBJ_DTYPE_OTHER, null, false);
+        $this->initVar('bcachetime', XOBJ_DTYPE_INT, 0, false);
+        $this->initVar('last_modified', XOBJ_DTYPE_INT, 0, false);
+
+		$xoops = \Xoops::getInstance();
+		
+        // for backward compatibility
+        if (isset($id)) {
+            if (is_array($id)) {
+                $this->assignVars($id);
+            } else {
+                $blkhandler = $xoops->getHandlerBlock();
+                $obj = $blkhandler->get($id);
+                foreach (array_keys($obj->getVars()) as $i) {
+                    $this->assignVar($i, $obj->getVar($i, 'n'));
+                }
+            }
+        }
+		$this->xoops_url = \XoopsBaseConfig::get('url');
+		$this->xoops_root_path = \XoopsBaseConfig::get('root-path');
+    }
+
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function id($format = 'n')
+    {
+        return $this->getVar('bid', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function bid($format = '')
+    {
+        return $this->getVar('bid', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function mid($format = '')
+    {
+        return $this->getVar('mid', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function func_num($format = '')
+    {
+        return $this->getVar('func_num', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function options($format = '')
+    {
+        return $this->getVar('options', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function name($format = '')
+    {
+        return $this->getVar('name', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function title($format = '')
+    {
+        return $this->getVar('title', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function content($format = '')
+    {
+        return $this->getVar('content', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function side($format = '')
+    {
+        return $this->getVar('side', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function weight($format = '')
+    {
+        return $this->getVar('weight', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function visible($format = '')
+    {
+        return $this->getVar('visible', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function block_type($format = '')
+    {
+        return $this->getVar('block_type', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function c_type($format = '')
+    {
+        return $this->getVar('c_type', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function isactive($format = '')
+    {
+        return $this->getVar('isactive', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function dirname($format = '')
+    {
+        return $this->getVar('dirname', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function func_file($format = '')
+    {
+        return $this->getVar('func_file', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function show_func($format = '')
+    {
+        return $this->getVar('show_func', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function edit_func($format = '')
+    {
+        return $this->getVar('edit_func', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function template($format = '')
+    {
+        return $this->getVar('template', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function bcachetime($format = '')
+    {
+        return $this->getVar('bcachetime', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function last_modified($format = '')
+    {
+        return $this->getVar('last_modified', $format);
+    }
+
+    /**
+     * return the content of the block for output
+     *
+     * @param string $format
+     * @param string $c_type type of content<br>
+     * Legal value for the type of content<br>
+     * <ul><li>H : custom HTML block
+     * <li>P : custom PHP block
+     * <li>S : use text sanitizater (smilies enabled)
+     * <li>T : use text sanitizater (smilies disabled)</ul>
+     * @return string content for output
+     */
+    public function getContent($format = 's', $c_type = 'T')
+    {
+        $format = strtolower($format);
+        $c_type = strtoupper($c_type);
+        switch ($format) {
+            case 's':
+                // check the type of content
+                // H : custom HTML block
+                // P : custom PHP block
+                // S : use text sanitizater (smilies enabled)
+                // T : use text sanitizater (smilies disabled)
+                if ($c_type == 'H') {
+                    return str_replace('{X_SITEURL}', $this->xoops_url . '/', $this->getVar('content', 'n'));
+                } else {
+                    if ($c_type == 'P') {
+                        ob_start();
+                        echo eval($this->getVar('content', 'n'));
+                        $content = ob_get_contents();
+                        ob_end_clean();
+                        return str_replace('{X_SITEURL}', $this->xoops_url . '/', $content);
+                    } else {
+                        if ($c_type == 'S') {
+                            $myts = \MyTextSanitizer::getInstance();
+                            $content = str_replace('{X_SITEURL}', $this->xoops_url . '/', $this->getVar('content', 'n'));
+                            return $myts->displayTarea($content, 1, 1);
+                        } else {
+                            $myts = \MyTextSanitizer::getInstance();
+                            $content = str_replace('{X_SITEURL}', \XoopsBaseConfig::get('url') . '/', $this->getVar('content', 'n'));
+                            return $myts->displayTarea($content, 1, 0);
+                        }
+                    }
+                }
+                break;
+            case 'e':
+                return $this->getVar('content', 'e');
+                break;
+            default:
+                return $this->getVar('content', 'n');
+                break;
+        }
+    }
+
+    /**
+     * (HTML-) form for setting the options of the block
+     *
+     * @return string|false HTML for the form, FALSE if not defined for this block
+     */
+    public function getOptions()
+    {
+        $xoops = \Xoops::getInstance();
+        if (!$this->isCustom()) {
+            $edit_func = (string)$this->getVar('edit_func');
+            if (!$edit_func) {
+                return false;
+            }
+            if (\XoopsLoad::fileExists($this->xoops_root_path . '/modules/' . $this->getVar('dirname') . '/blocks/' . $this->getVar('func_file'))) {
+                $xoops->loadLanguage('blocks', $this->getVar('dirname'));
+                include_once $this->xoops_root_path . '/modules/' . $this->getVar('dirname') . '/blocks/' . $this->getVar('func_file');
+                if (function_exists($edit_func)) {
+                    // execute the function
+                    $options = explode('|', $this->getVar('options'));
+                    $edit_form = $edit_func($options);
+                    if (!$edit_form) {
+                        return false;
+                    }
+                } else {
+                    return false;
+                }
+                return $edit_form;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+
+    public function isCustom()
+    {
+        return in_array($this->getVar("block_type"), array('C', 'E'));
+    }
+
+    /************ADDED**************/
+
+
+    /**
+     * Build Block
+     *
+     * @return array|bool
+     */
+    public function buildBlock()
+    {
+        $xoops = \Xoops::getInstance();
+        $block = array();
+        if (!$this->isCustom()) {
+            // get block display function
+            $show_func = (string)$this->getVar('show_func');
+            if (!$show_func) {
+                return false;
+            }
+            if (!\XoopsLoad::fileExists($func_file = $xoops->path('modules/' . $this->getVar('dirname') . '/blocks/' . $this->getVar('func_file')))) {
+                return false;
+            }
+            // must get lang files b4 including the file
+            // some modules require it for code that is outside the function
+            $xoops->loadLanguage('blocks', $this->getVar('dirname'));
+            $xoops->loadLocale($this->getVar('dirname'));
+            include_once $func_file;
+
+            if (function_exists($show_func)) {
+                // execute the function
+                $options = explode('|', $this->getVar('options'));
+                $block = $show_func($options);
+                if (!$block) {
+                    return false;
+                }
+            } else {
+                return false;
+            }
+        } else {
+            // it is a custom block, so just return the contents
+            $block['content'] = $this->getContent('s', $this->getVar('c_type'));
+            if (empty($block['content'])) {
+                return false;
+            }
+        }
+        return $block;
+    }
+}
