+++ htdocs/xoops_lib/Xoops/Core/Kernel/Handlers/XoopsBlockmodulelinkHandler.php
@@ -1,0 +1,38 @@
+<?php
+/*
+ You may not change or alter any portion of this comment or credits
+ of supporting developers from this source code or any supporting source code
+ which is considered copyrighted (c) material of the original comment or credit authors.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+*/
+
+namespace Xoops\Core\Kernel\Handlers;
+
+use Xoops\Core\Database\Connection;
+use Xoops\Core\Kernel\XoopsPersistableObjectHandler;
+
+/**
+ * XOOPS Kernel Class
+ *
+ * @copyright       XOOPS Project (http://xoops.org)
+ * @license         GNU GPL 2 or later (http://www.gnu.org/licenses/gpl-2.0.html)
+ * @package         kernel
+ * @since           2.6.0
+ * @author          Gregory Mage (AKA Mage)
+ * @author          trabis <lusopoemas@gmail.com>
+ * @version         $Id$
+ */
+
+class XoopsBlockmodulelinkHandler extends XoopsPersistableObjectHandler
+{
+    /**
+     * @param null|Connection $db
+     */
+    public function __construct(Connection $db = null)
+    {
+        parent::__construct($db, 'block_module_link', 'XoopsBlockmodulelink', 'block_id', 'module_id');
+    }
+}
