+++ htdocs/xoops_lib/Xoops/Core/Kernel/Handlers/XoopsRanksHandler.php
@@ -1,0 +1,39 @@
+<?php
+/*
+ You may not change or alter any portion of this comment or credits
+ of supporting developers from this source code or any supporting source code
+ which is considered copyrighted (c) material of the original comment or credit authors.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+*/
+
+/**
+ * XOOPS Kernel Class
+ *
+ * @copyright       XOOPS Project (http://xoops.org)
+ * @license         GNU GPL 2 or later (http://www.gnu.org/licenses/gpl-2.0.html)
+ * @package         kernel
+ * @since           2.6.0
+ * @author          Gregory Mage (AKA Mage)
+ * @author          trabis <lusopoemas@gmail.com>
+ * @version         $Id$
+ */
+
+namespace Xoops\Core\Kernel\Handlers;
+
+use Xoops\Core\Database\Connection;
+use Xoops\Core\Kernel\XoopsPersistableObjectHandler;
+
+class XoopsRanksHandler extends XoopsPersistableObjectHandler
+{
+
+    /**
+     * @param Connection $db
+     */
+    public function __construct(Connection $db)
+    {
+        parent::__construct($db, 'ranks', 'XoopsRanks', 'rank_id', 'rank_title');
+    }
+}
