+++ htdocs/xoops_lib/Xoops/Core/Kernel/Handlers/XoopsTplset.php
@@ -1,0 +1,99 @@
+<?php
+/**
+ * XOOPS kernel class
+ *
+ * You may not change or alter any portion of this comment or credits
+ * of supporting developers from this source code or any supporting source code
+ * which is considered copyrighted (c) material of the original comment or credit authors.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * @copyright       XOOPS Project (http://xoops.org)
+ * @license         GNU GPL 2 or later (http://www.gnu.org/licenses/gpl-2.0.html)
+ * @package         kernel
+ * @since           2.0.0
+ * @author          Kazumi Ono (AKA onokazu) http://www.myweb.ne.jp/, http://jp.xoops.org/
+ * @version         $Id$
+ */
+
+namespace Xoops\Core\Kernel\Handlers;
+
+use Xoops\Core\Kernel\XoopsObject;
+
+/**
+ * A Template Set File
+ *
+ * @author Kazumi Ono <onokazu@xoops.org>
+ * @copyright copyright (c) 2000 XOOPS.org
+ * @package kernel
+ **/
+class XoopsTplset extends XoopsObject
+{
+
+    /**
+     * constructor
+     **/
+    public function __construct()
+    {
+        $this->initVar('tplset_id', XOBJ_DTYPE_INT, null, false);
+        $this->initVar('tplset_name', XOBJ_DTYPE_OTHER, null, false);
+        $this->initVar('tplset_desc', XOBJ_DTYPE_TXTBOX, null, false, 255);
+        $this->initVar('tplset_credits', XOBJ_DTYPE_TXTAREA, null, false);
+        $this->initVar('tplset_created', XOBJ_DTYPE_INT, 0, false);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function id($format = 'n')
+    {
+        return $this->getVar('tplset_id', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function tplset_id($format = '')
+    {
+        return $this->getVar('tplset_id', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function tplset_name($format = '')
+    {
+        return $this->getVar('tplset_name', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function tplset_desc($format = '')
+    {
+        return $this->getVar('tplset_desc', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function tplset_credits($format = '')
+    {
+        return $this->getVar('tplset_credits', $format);
+    }
+
+    /**
+     * @param string $format
+     * @return mixed
+     */
+    public function tplset_created($format = '')
+    {
+        return $this->getVar('tplset_created', $format);
+    }
+}
